requirements.txt

aiofiles==23.2.1
aiohttp==3.9.1
APScheduler==3.10.4
asyncio==3.4.3
cachetools==5.3.3
dateparser==1.2.0
future==1.0.0
html2text==2020.1.16
httpx[http2]
motor==3.3.2
pymongo[srv]==4.6.2
pyrate-limiter==3.1.1
pytz==2024.1
python-arq==6.0.5
python-dotenv
pyrogram==2.0.106
python-telegram-bot==20.8
requests==2.31.0
telegraph==2.2.0
telethon==1.34.0
tgcrypto==1.2.5
